refactor: move react tx props type
